+++ examples/calendar/demo.py
@@ -7,8 +7,8 @@ SCHOOL_ID: int = 0
 
 
 def main():
-    server = api.Server(AUTH, USER_ID, SCHOOL_ID)
-    calender = server.get_calendar()  # get the student's personal calender from today
+    client = api.Client(AUTH, USER_ID, SCHOOL_ID)
+    calender = client.get_calendar()  # get the student's personal calender from today
     for (
         ctask
     ) in (
@@ -21,10 +21,10 @@ def main():
 
 def public_calender():
     # The entire school's calender
-    # It's the same code as above but with a different server function
+    # It's the same code as above but with a different client function
 
-    server = api.Server(AUTH, USER_ID, SCHOOL_ID)
-    calender = server.get_school_calendar()  # < The different server function
+    client = api.Client(AUTH, USER_ID, SCHOOL_ID)
+    calender = client.get_school_calendar()  # <-- The different client function
     for ctask in calender:
         print(
             f"{ctask.id} | Subject: {ctask.subject} | Teacher: {ctask.teacher_name} | Name: {ctask.title} | Due on: {ctask.due_on}"