+++ CHANGELOG.md
@@ -3,3 +3,11 @@
 ## Version 0.1 (development)
 
 - Initial Code
+
+## Version 0.2 (development)
+
+- BREAKING CHANGE: Rename api.Server class to api.Client
+- Add support to changing client ids (switch between Web and Android)
+- Fix login demo
+- Add Attendance
+- Add Detentions